pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gridzero"
version = "0.1.0"
description = "Transformer + GSPO for autonomous power grid control via grid2op"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "grid2op>=1.10",
    "lightsim2grid",
    "torch>=2.3",
    "transformers>=4.45",
    "vllm>=0.5",
    "outlines>=0.1",
    "pydantic>=2.0",
    "numpy",
    "gymnasium",
    "hydra-core>=1.3",
    "omegaconf",
    "torch-geometric",
    "aiohttp",
    "tqdm",
    "pytest",
    "pytest-asyncio",
]

[project.optional-dependencies]
dev = ["black", "ruff", "pytest", "pytest-asyncio"]

[tool.hatch.build.targets.wheel]
packages = ["gridzero"]

[tool.ruff]
line-length = 100
target-version = "py310"
